# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Video Editor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description?: string;
  /** JSON stringified editor state */
  state: string;
  /** Total duration in seconds */
  duration: number;
  canvasWidth: number;
  canvasHeight: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  description?: string;
  canvasWidth?: number;
  canvasHeight?: number;
}

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  state?: string;
  duration?: number;
  canvasWidth?: number;
  canvasHeight?: number;
}

export interface ProcessInstructionBody {
  /** Natural language instruction from the user */
  instruction: string;
  /** JSON stringified current editor state for context */
  currentState?: string;
}

/**
 * Operation-specific payload
 */
export type EditorOperationPayload = { [key: string]: unknown };

export interface EditorOperation {
  /** Operation type (e.g. addClip, cutClip, cropClip, moveClip, setAnimation, addTransition, setOpacity, setVolume, deleteClip) */
  type: string;
  /** Operation-specific payload */
  payload: EditorOperationPayload;
}

export interface InstructionResult {
  operations: EditorOperation[];
  /** Human-readable explanation of what will be done */
  explanation: string;
  /** Confidence score 0-1 */
  confidence: number;
}