# lib/api-zod/src/generated/api/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Video Editor API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  state: zod.string().describe("JSON stringified editor state"),
  duration: zod.number().describe("Total duration in seconds"),
  canvasWidth: zod.number(),
  canvasHeight: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const createProjectBodyCanvasWidthDefault = 1920;
export const createProjectBodyCanvasHeightDefault = 1080;

export const CreateProjectBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  canvasWidth: zod.number().default(createProjectBodyCanvasWidthDefault),
  canvasHeight: zod.number().default(createProjectBodyCanvasHeightDefault),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  state: zod.string().describe("JSON stringified editor state"),
  duration: zod.number().describe("Total duration in seconds"),
  canvasWidth: zod.number(),
  canvasHeight: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  state: zod.string().optional(),
  duration: zod.number().optional(),
  canvasWidth: zod.number().optional(),
  canvasHeight: zod.number().optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  state: zod.string().describe("JSON stringified editor state"),
  duration: zod.number().describe("Total duration in seconds"),
  canvasWidth: zod.number(),
  canvasHeight: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Process a natural language instruction into editor operations
 */
export const ProcessInstructionBody = zod.object({
  instruction: zod
    .string()
    .describe("Natural language instruction from the user"),
  currentState: zod
    .string()
    .optional()
    .describe("JSON stringified current editor state for context"),
});

export const ProcessInstructionResponse = zod.object({
  operations: zod.array(
    zod.object({
      type: zod
        .string()
        .describe(
          "Operation type (e.g. addClip, cutClip, cropClip, moveClip, setAnimation, addTransition, setOpacity, setVolume, deleteClip)",
        ),
      payload: zod
        .record(zod.string(), zod.unknown())
        .describe("Operation-specific payload"),
    }),
  ),
  explanation: zod
    .string()
    .describe("Human-readable explanation of what will be done"),
  confidence: zod.number().describe("Confidence score 0-1"),
});